+++ NEWS.md
@@ -61,6 +61,7 @@
   seconds) via the new `time` keyword argument.
 - A one-sided (or no-sided) taper can be applied using the new keywords
   arguments to `taper[!]` `left` and `right`.
+- One-sided tapers can be applied to up to 100% of the trace length.
 
 ## Deprecated or removed
 - `traces_are_orthogonal` will be removed in v0.5.  It has been replaced by